fix: allow passed table_schema in information_schema queries to vttablet if the keyspace is not found
