--- go/test/endtoend/vtgate/dbnameoverride/tablet_master_test.go
@@ -1,134 +1,0 @@
-/*
-Copyright 2019 The Vitess Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package master
-
-import (
-	"context"
-	"flag"
-	"os"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	"vitess.io/vitess/go/mysql"
-
-	"vitess.io/vitess/go/test/endtoend/cluster"
-)
-
-var (
-	clusterInstance *cluster.LocalProcessCluster
-	vtParams        mysql.ConnParams
-	hostname        = "localhost"
-	keyspaceName    = "ks"
-	cell            = "zone1"
-	sqlSchema       = `
-	create table t1(
-		id bigint,
-		value varchar(16),
-		primary key(id)
-	) Engine=InnoDB;
-`
-
-	vSchema = `
-	{
-    "sharded": true,
-    "vindexes": {
-      "hash": {
-        "type": "hash"
-      }
-    },
-    "tables": {
-      "t1": {
-        "column_vindexes": [
-          {
-            "column": "id",
-            "name": "hash"
-          }
-        ]
-      }
-    }
-  }`
-)
-
-const dbName = "myDbName"
-
-func TestMain(m *testing.M) {
-	defer cluster.PanicHandler(nil)
-	flag.Parse()
-
-	exitCode := func() int {
-		clusterInstance = cluster.NewCluster(cell, hostname)
-		defer clusterInstance.Teardown()
-
-		// Start topo server
-		err := clusterInstance.StartTopo()
-		if err != nil {
-			return 1
-		}
-
-		// Set extra tablet args for lock timeout
-		clusterInstance.VtTabletExtraArgs = []string{
-			"-init_db_name_override", dbName,
-		}
-
-		// Start keyspace
-		keyspace := &cluster.Keyspace{
-			Name:      keyspaceName,
-			SchemaSQL: sqlSchema,
-			VSchema:   vSchema,
-		}
-
-		if err = clusterInstance.StartUnshardedKeyspace(*keyspace, 1, false); err != nil {
-			return 1
-		}
-
-		if err = clusterInstance.StartVtgate(); err != nil {
-			return 1
-		}
-		vtParams = mysql.ConnParams{
-			Host: clusterInstance.Hostname,
-			Port: clusterInstance.VtgateMySQLPort,
-		}
-
-		return m.Run()
-	}()
-	os.Exit(exitCode)
-}
-
-func TestDbNameOverride(t *testing.T) {
-	defer cluster.PanicHandler(t)
-	ctx := context.Background()
-	conn, err := mysql.Connect(ctx, &vtParams)
-	require.Nil(t, err)
-	defer conn.Close()
-	qr, err := conn.ExecuteFetch("SELECT database() FROM information_schema.tables WHERE table_schema = database()", 1000, true)
-
-	require.Nil(t, err)
-	require.Equal(t, 1, len(qr.Rows), "did not get enough rows back")
-	require.Equal(t, dbName, qr.Rows[0][0].ToString())
-}
-
-func TestInformationSchemaQuery(t *testing.T) {
-	defer cluster.PanicHandler(t)
-	ctx := context.Background()
-	conn, err := mysql.Connect(ctx, &vtParams)
-	require.Nil(t, err)
-	defer conn.Close()
-	qr, err := conn.ExecuteFetch("SELECT TABLE_NAME FROM information_schema.tables WHERE table_schema = 'ks'", 1000, true)
-
-	require.Nil(t, err)
-	require.Equal(t, 1, len(qr.Rows), "did not get enough rows back")
-	require.Equal(t, "t1", qr.Rows[0][0].ToString())
-}
